+++ offline/packages/intt/InttVertexUtil.h
@@ -1,8 +1,7 @@
 #ifndef INTTVERTEXUTIL_H
 #define INTTVERTEXUTIL_H
 
-#include <TF1.h>
-#include <TMath.h>
+#include <vector>
 
 namespace InttVertexUtil {
   double gaus_func(double *x, double *par);
